Split nn_syntax module into multiple modules
Makes it easier to understand what's going on
