+++ Makefile
@@ -5,16 +5,16 @@
 %.js : %.pegjs
 	node ./node_modules/.bin/pegjs -o $@ $<
 
-lib/grammar.js : lib/grammar.pegjs
-	node ./node_modules/.bin/pegjs --allowed-start-rules input,program,type_ref,permission_rule -o $@ $<
-
 %.mo : %.po
 	msgfmt $< -o $@
 
 all = \
-	$(patsubst %.lr,%.js,$(wildcard lib/syntax/*.lr)) \
 	$(patsubst %.po,%.mo,$(wildcard po/*.po)) \
+	lib/nn-syntax/parser.js \
 	lib/grammar.js \
 	test/test_sr_parser_generator.js
 
 all : $(all)
+
+lib/grammar.js : lib/grammar.pegjs
+	node ./node_modules/.bin/pegjs --allowed-start-rules input,program,type_ref,permission_rule -o $@ $<
