+++ lib/nn-syntax/parser.lr
@@ -163,10 +163,10 @@ rule = {
 }
 
 aggr_op = {
-    'min';
-    'max';
-    'sum';
-    'avg';
+    'min' => 'min';
+    'max' => 'max';
+    'sum' => 'sum';
+    'avg' => 'avg';
 }
 
 table = {
@@ -181,9 +181,9 @@ table = {
     'aggregate' 'argmin' field:out_param base:constant_Number ',' limit:constant_Number 'of' '(' table ')' => new Ast.Table.ArgMinMax(table, field.name, op, base, limit, null);
     'aggregate' 'argmax' field:out_param base:constant_Number ',' limit:constant_Number 'of' '(' table ')' => new Ast.Table.ArgMinMax(table, field.name, op, base, limit, null);
     'window' base:constant_Number ',' delta:constant_Number 'of' '(' stream ')'        => new Ast.Table.Window(base, delta, stream, null);
-    'timeseries' base:constant_Date ',' delta:constant_Measure(ms) 'of' '(' stream ')' => new Ast.Table.TimeSeries(base, delta, stream, null);
+    'timeseries' base:constant_Date ',' delta:constant_Measure_ms 'of' '(' stream ')' => new Ast.Table.TimeSeries(base, delta, stream, null);
     'sequence' base:constant_Number ',' delta:constant_Number 'of' '(' table ')'       => new Ast.Table.Sequence(base, delta, table, null);
-    'history' base:constant_Date ',' delta:constant_Measure(ms) 'of' '(' table ')'     => new Ast.Table.History(base, delta, table, null);
+    'history' base:constant_Date ',' delta:constant_Measure_ms 'of' '(' table ')'     => new Ast.Table.History(base, delta, table, null);
 
     table_join;
 }
