+++ lib/nn-syntax/tonn_converter.js
@@ -2,190 +2,21 @@
 //
 // This file is part of ThingTalk
 //
-// Copyright 2017 The Board of Trustees of the Leland Stanford Junior University
+// Copyright 2017-2018 The Board of Trustees of the Leland Stanford Junior University
 //
 // Author: Giovanni Campagna <gcampagn@cs.stanford.edu>
 //
 // See COPYING for details
 "use strict";
 
-const adt = require('adt');
 const util = require('util');
 
-const Ast = require('./ast');
-const Type = require('./type');
-const NNOutputParser = require('./syntax/thingtalk');
-const { parseDate } = require('./date_utils');
+const Ast = require('../ast');
+const Type = require('../type');
+const { parseDate } = require('../date_utils');
 
-class TokenWrapper {
-    constructor(token, value, location) {
-        this.token = token;
-        this.value = value;
-        this.location = location;
-    }
-
-    toString() {
-        return this.token;
-    }
-}
-
-class SequenceLexer {
-    constructor(sequence, entities) {
-        this._sequence = sequence;
-        if (!Array.isArray(sequence))
-            this._sequence = Array.from(sequence);
-
-        if (typeof entities !== 'function') {
-            this._entities = (next) => {
-                if (!(next in entities)) {
-                    if (next.startsWith('SLOT_'))
-                        return undefined;
-                    throw new SyntaxError('Invalid entity ' + next + ', have ' + Object.keys(entities));
-                }
-                return entities[next];
-            };
-        } else {
-            this._entities = entities;
-        }
-
-        this._i = 0;
-        this._lastfunction = null;
-        this._lastparam = null;
-        this._instring = false;
-    }
-
-    next() {
-        if (this._i >= this._sequence.length)
-            return { done: true };
-
-        let next = this._sequence[this._i++];
-        if (next === '"') {
-            this._instring = !this._instring;
-        } else if (this._instring) {
-            next = new TokenWrapper('WORD', next, i);
-        } else if (/^[A-Z]/.test(next)) {
-            // check if we have a unit next, to pass to the entity retriever
-            let unit = null;
-            // note that this._i has already been increased
-            if (this._i < this._sequence.length && this._sequence[this._i].startsWith('unit:'))
-                unit = this._sequence[this._i].substring('unit:'.length);
-
-            // entity
-            const entity = this._entities(next, this._lastparam, this._lastfunction, unit);
-            const entityType = next.substring(0, next.lastIndexOf('_'));
-            if (entityType.startsWith('GENERIC_ENTITY_')) {
-                next = new TokenWrapper('GENERIC_ENTITY', {
-                    value: entity.value,
-                    display: entity.display,
-                    type: entityType.substring('GENERIC_ENTITY_'.length)
-                });
-            } else {
-                next = new TokenWrapper(entityType, entity);
-            }
-        } else if (next.startsWith('@')) {
-            this._lastfunction = next;
-            let lastPeriod = next.lastIndexOf('.');
-            let kind = next.substring(1, lastPeriod);
-            let channel = next.substring(lastPeriod+1);
-            if (!kind || !channel)
-                throw new Error('Invalid function ' + next);
-            if (channel === '*')
-                next = new TokenWrapper('CLASS_STAR', kind);
-            else
-                next = new TokenWrapper('FUNCTION', { kind, channel });
-        } else if (next.startsWith('enum:')) {
-            next = new TokenWrapper('ENUM', next.substring('enum:'.length));
-        } else if (next.startsWith('param:')) {
-            let [,paramname,] = next.split(':');
-            this._lastparam = paramname;
-            next = new TokenWrapper('PARAM_NAME', paramname);
-        } else if (next.startsWith('unit:')) {
-            next = new TokenWrapper('UNIT', next.substring('unit:'.length));
-        } else if (next.startsWith('device:')) {
-            next = new TokenWrapper('DEVICE', next.substring('device:'.length));
-        } else if (next.startsWith('^^')) {
-            next = new TokenWrapper('ENTITY_TYPE', next.substring('^^'.length));
-        }
-        return { done: false, value: next };
-    }
-}
-
-
-function fromNN(sequence, entities) {
-    let parser = new NNOutputParser();
-    return parser.parse({
-        [Symbol.iterator]() {
-            return new SequenceLexer(sequence, entities);
-        }
-    });
-}
-
-
-// A lazy functional list
-const List = adt.data(function() {
-    return {
-        Nil: null,
-        Cons: {
-            head: adt.any,
-            tail: adt.only(this)
-        },
-        Snoc: {
-            head: adt.only(this),
-            tail: adt.any
-        },
-        Concat: {
-            first: adt.only(this),
-            second: adt.only(this)
-        }
-    };
-});
-List.prototype.flatten = function(into) {
-    if (this.isNil)
-        return into;
-    if (this.isCons) {
-        into.push(this.head);
-        return this.tail.flatten(into);
-    } else if (this.isSnoc) {
-        this.head.flatten(into);
-        into.push(this.tail);
-        return into;
-    } else if (this.isConcat) {
-        this.first.flatten(into);
-        return this.second.flatten(into);
-    } else {
-        throw new TypeError();
-    }
-};
-List.prototype.getFirst = function() {
-    if (this.isNil)
-        return null;
-    if (this.isCons)
-        return this.head;
-    if (this.isSnoc)
-        return this.head.getFirst();
-    if (this.isConcat)
-        return this.first.getFirst();
-    throw new TypeError();
-};
-List.concat = function(...lists) {
-    let result = List.Nil;
-    for (let i = lists.length-1; i >= 0; i--) {
-        if (lists[i] instanceof List)
-            result = List.Concat(lists[i], result);
-        else
-            result = List.Cons(lists[i], result);
-    }
-    return result;
-};
-List.singleton = function(el) {
-    return List.Cons(el, List.Nil);
-};
-
-class UnsynthesizableError extends Error {
-    constructor(what) {
-        super(what + ' cannot be synthesized');
-    }
-}
+const List = require('./list');
+const { UnsynthesizableError } = require('./errors');
 
 function unescape(symbol) {
     return symbol.replace(/_([0-9a-fA-Z]{2}|_)/g, (match, ch) => {
@@ -262,7 +93,7 @@ function filterToCNF(filter) {
     return Ast.BooleanExpression.And(clauses);
 }
 
-class ToNNConverter {
+module.exports = class ToNNConverter {
     constructor(sentence, entities) {
         if (typeof sentence === 'string')
             sentence = sentence.split(' ');
@@ -822,21 +653,4 @@ class ToNNConverter {
         else
             throw new TypeError();
     }
-}
-
-function toNN(program, sentence, entities) {
-    // for backward compatibility with the old API
-    if (!entities) {
-        entities = sentence;
-        sentence = '';
-    }
-
-    let converter = new ToNNConverter(sentence, entities);
-    return converter.toNN(program);
-}
-
-module.exports = {
-    fromNN,
-    toNN,
-    UnsynthesizableError
 };
